Create a new folder, test_symbolic, with and emply file. In symbolic class create christoffel, tensor, riemann class. And create others files (as ricci, metric...), which are emply
