+++ src/einsteinpy/symbolic/einstein_symbolic.py
@@ -1,0 +1,4 @@
+from einsteinpy.symbolic import Tensor
+
+class EinsteinTensor(Tensor):
+    pass
