+++ src/einsteinpy/symbolic/metric_symbolic.py
@@ -1,0 +1,4 @@
+from einsteinpy.symbolic import Tensor
+
+class MetricTensor(Tensor):
+    pass
