+++ src/einsteinpy/symbolic/ricci_symbolic.py
@@ -1,0 +1,4 @@
+from einsteinpy.symbolic import Tensor
+
+class RicciTensor(Tensor):
+    pass
