+++ src/einsteinpy/symbolic/stress_energy_momentum_symbolic.py
@@ -1,0 +1,4 @@
+from einsteinpy.symbolic import Tensor
+
+class StressEnergyMomentumTensor(Tensor):
+    pass
